pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bitswan-on-prem-aoc"
version = "0.1.0"
description = "Command line tool for AOC environment setup"
requires-python = ">=3.10"
authors = [{ name = "LibertyAces Ltd.", email = "support@bitswan.space" }]
dependencies = [
  "PyYAML==6.0.2",
  "python-keycloak==5.1.1",
  "requests==2.32.3",
  "influxdb-client==1.48.0",
  "click==8.1.8",
  "python-dotenv==1.0.1",
  "aiohttp==3.11.13",
]

[project.scripts]
bitswan-on-prem-aoc = "aoc_cli.commands.cli:cli"


[tool.hatch.build.targets.wheel]
packages = ["aoc_cli"]
